Finished writing changes to JSON file on local machine
Still needs testing with game - currently tabs in original JSON are
replaced with 2 spaces. Results not yet formally checked, but new JSON
file is the expected length i.e. don't appear to have broken the item
files.

Also wrote the code allowing the xy_changes only option to work - when
want to keep the quoin types as they currently are on the street.