Sanity checking new JSON files - mostly done
All item JSONs are saved in data/new - so can be validated before
writing to persdata. However it looks like doing a character count of
the changes and then comparing it to the final file length doesn't work
because sometimes a field is an int, other times a string - and then the
maths breakdown. So will assume the json-writing is working fine, and
whilst in testing mode, run my json-comparison tool from within this
tool to report on the differences between original/new json. This will
be turned off once the tool is made available to others. This work will
be part of the next commit.