Replaced JSON diff mechanism
Originally I was counting the number of characters in both old/new JSONs
and seeing if the difference matched what I was expecting. However some
fields are can be either strings/ints - which have a different number of
characters in the file, so can't be compared that way. This is a proper
JSON diff tool which walks down the keys comparing/finding keys. It is
assumed that keys will be either changed/added to in the new JSON -
deletion of a key is an error and handled as such.