Fixed various bugs
Tool now handles missing/skipped items, multiple snaps. Still some
memory leakage - but this happens if just the top level is looped over
(comment out the call to ProcessItem()) - made worse by tracing. So
hopefully won't be too much of a problem in the final tool? Just done
GFJ with 60 items - the only fail was that it failed to locate the
shrine which seems odd. But was able to skip the street spirit and
handle searching for 58 trees/quoins.