Reduce item image to black/white to compensate for tinted/contrast etc in street G* file
Didn't realise that the items on a street will have their colour changed
depending on the contrast/tint/brightness/saturation in the street G*
file. As a first hack the item/street fragment are converted to
black/white mode and these 2 images are compared. Works better than if
simple file comparisons are done. However it still might be necessary to
apply the street G* settings to each item fragment on that street, if
this black/white hack is not reliable at finding items.

Fixed some minor bugs - but still need to clean up handing
skipped/finished/found but still searching flags as problems caused when
skipped items such as street vendor mean the next item (a quoin) is not
searched for because the item has not been properly initialised.