Only use street snaps that match the h/w size in geo file
Snaps that are smaller than this don't give a useful search result if
enlarged up to the right pixel resolution. So just drop them if they
don't match the h/w in the geo file. Also if snaps/streets are invalid,
then flag up a warning for 2s but then continue on - so that the tool
will be able to work through a region without needing the user to hit
'c' to continue. If streets are skipped then the list is displayed on
the screen at the end (rather than closing the application automatically
for successful run).