Tool now prompts for location of config.json on initial run
In configLocation.txt in the programme directory, the location of the
config.json (and where the JSONs will be stored temporarily) will be
stored. If this file is missing, or the json file does not exist, then
the user will be prompted for the correct file location.

Also added extra 'silent' parameter to the ls sftp command so that I can
test to see if files exist in a directory remotely e.g. to test if
persdata exists or not.