Finished sftp upload./download
All the JSON files are downloaded together at the start of the program -
and then the connection is ended. After processing of all streets has
completed, all changed files are uploaded together, as part of a fresh
session.

All JSON files are copied to OrigJSONs directory in the same directory
as the config.json. Any item files that are changed are written to
NewJSONs. These are then copied to the vagrant directory/uploaded to the
server and if successful the file is then moved to UploadedJSONs.
Also, keep the contents of the working directories if debug flag set.

Known problems - currently uploading all the items per street - so if
the program is exited by the user, the sftp thread will continue running
until that street is finished. Need to find a better way of catching a
user closure of the screen. Or return to the top level after each sftp
file put/get. Which will need a reorganisation of the code.