Replaced 'append to file' option with automatic renaming of existing output file
If an output file already exists, then it is renamed to outputfile_x.txt
where x is the number of files that start with the outputfile name. So
if outputFile and outputFile_1 exist, then when the program is run,
outputFile will be moved to outputFile_2. Problems will happen if the
user then deletes outputFile_1 ... so the program will rename outputFile
to outputFile2 which already exists ... and so will fail. However the
code which returns the list of outputfiles puts them in alphabetical
order which means outputFile22 is earlier than outputFile8 so I can't
rely on outputFile22 being at the end of the list (and then simply add 1
to provide the next file version).