Handled quoins with same final x,y; printing fatal errors to screen
Sometimes a group of closely spaced quoins will be determined by the
program to have the same x,y - it isn't possible to work out whether
quoin A is actually the 2nd from the LHS in the cluster, or the one in
the middle. So the search results can be confusing. Works fine when a
quoin is either present/absent rather than a couple being close by to
the original x,y. So if this is found to be the case, the co-located
quoins are all treated as if they had not been found in the first place,
and it will be up to the user to compare to a snap to set them up
properly.

Realised that the final release of this program will not have a debug
window to println to - so any fatal errors are printed out to the screen
so the user can see them. A message is also printed to the output file
telling the user to look in debug_info.txt for more information about
the failure.