Differentiate quoin/item search radius in config file and bug fixes
When quoins are close together is useful to have a smaller search radius
- so split out quoin/non-quoin search radiuses in the config.json file.
Recommended 50 for the non-quoins and 20 for quoins.

Fixed a bug which meant that if co-located quoins were being changed
back to mystery/missing quoins, the amended x,y values were not being
updated in the json file before it was rewritten to NewJSONs directry.
Fixed a bug which meant that the saveChangedJSONFlag was being
incorrectly set when a quoin had been found in the same x,y,type as
before. No functional problem, but made the output look odd as it was
marked as ** when in fact nothing had changed. Noticable when the tool
was run repeatedly
Fixed a minor info bug which meant that if mystery quoins were found the
first time the tool was run, subsequent re-runs of the tool meant that
these mystery quoins were being updated again, when in fact nothing had
changed.