For pairs of co-located quoins, furthest original x,y is set to 'missing'
For each pair of co-located quoins, the quoin which has the furthest
distance to the original x,y is assumed to be incorrect and so set to be
'mystery'.

The list of summary results is continually sorted until there are no
more sets of co-located quoins left - a situation which is possible when
3 or more are co-located. As a failsafe, if the loop has gone round 10
times, then on the 10th time, both members of the pair are treated as
missing. Should avoid ever risking writing co-located quoins to
persdata.