When initially searching for quoin, search using all quoin images so can find the nearest one
Need to distinguish between closely spaced quoins. So need to
differentiate between the XP quoin which is further away than the time
quoin - but which nonetheless appears to be a match.  Can only do this
by searching for all quoin images, saving the information and then going
through to find the nearest found image.

This has resolved the bug that I'd seen in which the number of mystery
quoins increased if the tool was rerun on a street. Also seeing far less
cases of co-located quoins and don't appear to need the separate
quoin_radius setting in the config.json. But I'll leave it in there for
now, but set to 50 by default as then single quoins, badly placed will
still be found.