First stab at fetching/putting JSON files single and other changes
Rather than fetching all item/street JSON files at start of program, it
now fetches only the JSON files needed for the street being processed.
Once the changes are made, the JSON files are written to persdata before
moving on to the next street. Means that when the server files are being
used, quitting the program will halt at the next upload/download rather
than continuing on.

Added in item png files that were missing - although have not yet added
the code to handle them in the code.

Program ends and leaves the final message on the screen - now need to
press ESC or x to exit