Streets in persdata-qa are skipped
Added the feature which means any street which is found to exist in
persdata-qa will be skipped. Means that this tool can't overwrite quoins
simply because the snaps are missing them. If a user does want to change
the co-ordinates of items then the xy_only option can be used (although
this appears to be completely broken).

As this will be release v0.3 - have turned off the reporting of failures
to delete the JSON files from NewJSONs directory as still haven't been
able to fix that.