Exploring average pixel difference, saving best outcomes per search
At the moment the code determines goodness of match by looking at the
lowest RGB diff for an entire fragment. Traced the related values of
lowest RGB diff/pixel and the average RGB diff/pixel of all searches.
Trying to get a handle on what constitutes a good enough match given
that some rocks with mist in front of them are almost good enough to
pass. But still need to differentiate completely missing items.