Fragment matching now done on average RGB diff/pixel
A match is considered to happen if the lowest average RGB diff/pixel :
total average RGB diff/pixel over all searches is less than 0.1 for a
90% match. The %match level is configured in the config.json and this
figure of 90% seems to be better than the previous method where I just
considered the lowest RGB diff for an entire fragment being less than
5000. Has managed to successfully catch rocks with mist in front of them
for example. QQ have 15% additional leeway to allow for their size
change as they bounce - so 90% for items relates to 75% match for QQs to
be found.

Also improved the output given to the user which indicates the %match
for each item - so can see the confidence with which the x,y has been
determined.

Also log street snaps which are skipped to the user output file -
otherwise isn't immediately clear why a QQ isn't found when you've seen
it clearly in a snap!