Prints out the largest delta x or y between found/original x,y
This is simply additional information to be dumped out for a while in
the output file to see if the 25px radius commonly used is adequate. So
far most matches are found within 5px, but some things such as visiting
stones are nearer 20px. So 25px might still be a good default value.
Tracing will be removed at some point