Added support for defaulting quoins fields in non-standard regions
Tool now correctly defaults the quoin fields for Ancestral Lands, party
spaces, quion hubs etc. Testing this showed some minor bugs in how
changes were being reported to the user - so these were fixed as well.

Rainbow Run - server needs a code change to set 'rainbow run' quoins to
be currants in order to match the one video we have. At the moment they
are set to mood quoins. So this tool will accept mood/current quoins in
a snap, and then default to be currants. Everything else is marked as
mystery quoin.