Saves the best matchInfo when searching for item in FragmentFind
Prior to this fix, matchInfo was overwritten after each return from
SpiralSearch - hence for trees which were not found, the 'best nearest
match' info given to the user related to the data saved for the last
image search (a wood tree). Now the code only saves the match info from
SpiralSearch if it is better than the match info already saved for that
item. Therefore now a user will see that an egg tree was not found, but
actually there is a reasonable x,y at 87% ... which they can then
manually input.