Can dump images of the best match for each item
Using the option debug_dump_bw_diff_images, debug images can be dumped
in BestMatchImages in the Tool directory. These do a comparison of the
relevant street fragment with the item fragment and for all pixels which
are white in one fragment and black in the other, the pixel is coloured
red. Should be easy therefore to see why an item isn't being found.