Validate changed JSON files to check match item class
Sanity check that only changing expected fields in a JSON file. If
something unexpected is found then a report is made to both debug and
output file so that it can be investigated and not overlooked.

Also moved MatchInfo class to separate tab.

New versions of fragments for jellisacs/ice nubs as did not produce 100%
match when run these streets containing the original samples on this
tool.