Handle patches in same way as trees, skip streets with no items
As patches may well contain trees in snaps, these items have been
treated in the same way as trees. Therefore when searching for a tree,
the list of images now includes patches. This means that if our item
JSON is a patch, the correct x,y will be determined from any tree
position on a snap.

Fixed bug which meant the tool did not move on to the next street if the
street had no items to process.