Fix dump of BW images, and minor improvements
This fix now dumps out the correct black/white images for street/item
fragments as well as the red marked diff version.

When an item is missing, the x,y and variant information for this
nearest match is also given - otherwise wouldn't know which kind of ice
knob related to this x,y.

Ice knobs are transparent and so the match is not as good as hoped for -
so the code now manually looks for 90% if the specified % match
required, similar to the way the QQ requires a 85% match of the
user-specified value.