Only show message for non-silent LS SFTP commands
The logic was wrong - which meant that when an ls was done to see if a
street existed in persdata-qa, the failed LS command 'no such file'
caused an error message to be printed even when a 'silent' ls had been
requested. This message is now not seen for this check to see if the
street has already been QA'd and is in persdata-qa.