Added in support for validation run
By setting a flag in the config.json, carries out a validation run using
a pre-defined list of streets which incorporate all items. In addition
JSONDiff output is put into the validation output file. This validation
run also uses as source JSON files in a special directory in the
validation directory. Therefore in the future the tool can be re-run to
check the output is the same - thereby checking that changes have not
broken anything.