Tweaks plus partial fix to memory leak problem
When loading certain streets e.g. Chavila Checker, the programme
exceeded 256 MB, even when just one snap loaded and only about 7 quoins.
Removing the cache for the image when it was unloaded appeared to fix
the problem
(https://github.com/processing/processing/issues/1391.html#issuecomment-13356835).
There is still some memory problem which I think is down to Processing
as a simple load street snap/unload street snap showed that after the
unload, the memory didn't return to the initial level before the first
load - some memory is not being released. Not sure I can do anything
about that other than set the memory bar for the programme a bit higher
e.g. to 300mb.

As it is useful, on level 1 error tracing, memory information is dumped
out when loading/unloading images - can then be put in a spreadsheet
(import, split at ,) and graphed so can see what is happening.

Also changed quoin output info to only give the class name - type is
irrelevant as included in class name information. And improved some
tracing for when start processing an item on a street.