This version = build v0.7
Preferences also set to 300mb memory to help reduce risk of memory
leakage causing a crash (because street snaps leave behind residual
memory usage when loaded and then unloaded). When the street structure
is destroyed, the memory is finally freed, but not before then. So not
an issue with the number of streets parsed, but the number of snaps per
street (potentially, as doesn't happen for all streets).