Correctly handle quoins when changing x_y only
Although the code was correctly only changing x,y in the quoin JSON
files, the output file stated that these quoins were being reset to
mystery when not found. This could be because missing from the snap, or
if QA tool not yet run, then likely that the defaulted quoin types
didn't match the quoins in the snap.