Fix another change_xy_only bug - affecting other items with types
When this option is set, we should only search the street snap with the
specified fragment type as indicated in the JSON file. And report the
item as MISSING if this match fails. Previous to this, the code looped
through all the barnacle images and so always found a match in the snap,
even though all the items were defaulted to type 1 (and so therefore
some should fail). Now these non-type 1 barnacles in the street snap are
reported as MISSING.