Further fixes for change_xy_only option
Wood trees still need to have all tree images searched against the
street snap - as no guarantee that the wood tree JSON file is reflected
in the snap. However the type of wood tree will not be changed.

Carry out a sanity check in the JSON Diff functionality to ensure that
the x,y fields are the only changes ever made to JSON files when this
option is selected.