Corrected debug message for failed snaps
The previous debug message stated that the snap had been failed because
it was too small. But we also reject snaps that are too large (rare, but
happens). So have changed the debug message so that it does not make the
assumption about relative snap size. The message output to the output
file does not differentiate the two failure cases and so does not need
changing.