Implemented colour tinting of item fragment using geo data
This reduces the risk of false positives on dim streets. The item
fragment can now be tinted using the same geo data which is applied to
the street. Unfortunately this alone cannot be used to do a colour
comparison because some of the items are transparent and take up some of
the background colour. Therefore a black/white comparison will still be
carried out on these fragments. This has improved reliability and
additional items are now being found on streets.

As part of this change, new item fragments were regenerated against a
background where r=g=b=128, and none of these contain any of these
background pixels.