Produce summary validation report that is easier to eyeball
Cleaned up some of the duplicate code in PrintToFile by creating a new
class OutputFile which does the actual writing/opening etc.

A validation run now produces 2 files - validation.txt which contains
all the details of the changes together with JSON diff information. Also
a validation_summary.txt which just lists the details of the changed
items, and a count - this is much easier to compare against an earlier
version.

Comparing a black/white vs. tintied/black/white run shows the results
are largely identical. There are a couple of differences between but
nothing concerning. So moving to the tinted fragments I think has
produced more accurate results even if that now means a few items are
treated as missing. But this is better than a false positive.