Fixed bug which reported incorrect 'best match' for missing quoin
When saveClosestQuoin() is called, if no quoins that meet the matching
criteria have been found, then should just keep the existing
bestMatchInfo. Otherwise, the search results for the last quoin (time)
will be reported, rather than the 'best' missing result. This is
unlikely to be useful as will probably depend on the background scenary,
but still worth reporting.