Major rework of handling search results and minor updates
The decision to stop searching when a good enough match was found turned
out to lead to false positives, especially when using a large search
radius. So the whole of FragFind was rewritten so that searches now
continue on until all images have been searched or a PERFECT match has
been found. The only exception for this is quoins (where need to search
on every snap to make sure that find as many quoins as possible,
prioritising those closest to the original x,y) and trees, where always
need to search for all possible tree images. Otherwise a 'good enough'
match caused by accidental background matching with a bean tree trunk
will overwrite a much better match for an egg tree.

Rather than maintaining a list of 'bestMatchInfo' elements, each item
will simply have the 'best so far' saved. This structure will be used to
work out the new variant/x/y information as needed and will also be
updated when a lower y-value is found for a quoin/QQ. The use of
MISSING_COORDS is not used any more to indicate whether an item has been
found or not.  These changes simplify the code structure and get rid of
some hokey code that was developed as I discovered more edge cases that
needed handling. Hopefully this is more robust now - and passes the
validation test (fixed one case compared to previous run).

Fixed up duplicate quoin handling so that more useful information was
given to the user. Works better with the new search handling code
implemented here.

Also added configutation option to turn on/off information about the
pixel distance between new/original item location.