Partial implementation of street image dump showing missing/found items
Will produce a street snap which has items painted on to the first
street snap, and with a square box showing whether a match was found or
not. Simply to make it quicker to spot if an item was found, by eye.

Also includes some of the more recent updated fragments which have
transparent pixels on. Not yet tested