Several bugs and new minor features
Fixed bug which meant that x_y only mode was not taking account of the
multiple images for trees, barnacles etc, but was still assuming there
was one single image for these items. So searches were failing to find
the item when present on a street.

Fixed a bug which mean that a null pointer was found when checking for
co-located quoins - when the quoin was a qurazy quoin. The duplicate
quoin check now only happens for all other kinds of quoin, not this
transient one which we are not analysing.

Added support for enchanted wood trees.

Items with multiple images such as trees will now have the search marked
as 'done' if a 99% match or better is found in order to reduce the
number of images needing to be compared. As the mature versions are
always searched first, a 99% match is probably as good as it is going to
get. This is configurable for me in case I need to play with the value
in the future.

Added new debug option which allows me to dump out a list of matches for
an item, above a specifed % in the config.json file. Should help show
when accuracy doesn't appear to be happening as expected.