Removed gardening vendor, added zutto vendor
As the gardening vendor wobbles around on his pole, there is no point
trying to work out the co-ordinates. So this item will now be skipped.

Added back the zutto vendor - search for a normal and flipped version of
the fragment. Although I'm still not seeing a brilliant match - and the
placement looks to be a couple of pixels off.