Added in boxes around found/missing items on street summary picture
This allows the user to quickly see what has been found or is missing on
a street together with the image used to locate it. For missing items, a
red box the size of the search radius is drawn together with a red dot
where the original x,y was set. A smaller red box at the top (or bottom)
contains the nearest match item fragment for information purposes.