Support for stages of rock, peat and dirt and minor display fix
Tool now searches for the different stages of rock, peat and dirt. Start
with the most mature/full versions and then works down towards empty.

Changed the display on screen so that the red box around the sample
being searched for is in the centre of the street fragment, rather than
the top LH corner being in the centre - which caused a problem for wide
item fragments.