Added flag for using only mature images in search
The use_mature_item_images_only flag is optional - if missing, this is
defaulted to false. Allows the user to decide if they only want to
search using the mature images, rather than all possible images for an
item (this is especially useful for trees which have 10 images each).
Information about the value of this flag is included at the start of the
output file.

For a validation run, this flag is defaulted to false.

Minor change to only print out memory information for image loading for
street images, not item images.