Added street actions for those streets in/outside persdata-qa
Created new structures in config.json which specifically state what
action to take for streets in persdata-qa and those that are not. This
means the user can choose to skip streets in persdata-qa, or just tweak
the x,y. If the structures are missing, then streets in perdata-qa will
be skipped, other streets will have x,y and variants changed (i.e. as
before).

Streets which are deliberately skipped will result in a 1s warning
displayed in black, streets which encounter an error e.g. missing JSON
file will result in a 5s warning message in red.

Minor change - each street includes a summary count of the number of
changed/missing items and a message reminding the user to redo /qasave
for streets in persdata-qa when changed files have been written to
persdata.