Added option to print summary count of changed/uploaded JSON files and minor bug fixes
When running the tool, can get a long list of changed JSON files in the
output file. So user can turn this off if they want - and just get a
summary count of the files changed. If the field is missing from the
JSON file, then the full list of uploaded/copied fields is printed - as
for current behaviour.

Put a bigger message at the end of the street output if changes have
been made to streets in persdata-qa - to remind the user to redo /qasave