Add option for directory for street summary images
By default the street images showing the found/missing objects are
placed in the tool working directory - and this is cleared each time the
tool is run. However this means that it is easy to lose all the images
for a region run, if one then redoes the tool for a single street. The
directory specified in the option will not be cleared each time the tool
is run - it will be left up to the user to maintain. If the option is
missing, then the default directory will be used, as per the exisiting
code.