Changed handling of fields indicating how to handled persdata/persdata-qa streets
Currently the code assumed that if one/both of these structures were
present, then all 3 fields would be contained therein. But actually
there only needs to be the one which is set to 'true' - doesn't matter
if the other options which would be set to 'false' are missing. So the
code now handles this way of doing things.

Changed name of field in config file - and variables referring to this
field.