Separate out subterranean from above ground tree images
Realised that egg trees will only ever be found in dark patches.
Therefore if we have an egg tree in QA2, then there is no point also
checking for bean tree images etc. And vice versa - if we have a bean
tree, it could match a fruit tree image, but will not match an egg tree.
This has shaved about 10% off the run time of a street, so should be
useful.