Correctly handles case when user hits cancel when selecting config file location
This was causing a null pointer error because I was printing out debug
information before the structure had been set up. I have cleaned up wth
handling of this user dialog so that if the user just hits cancel, then
a message is reported to the screen and x/q/ESC can be used to quit.

Have also added in defensive code to only attempt to write to a file, if
the structure pointer is not NULL. Which should prevent anything like
this happening again.

Also updated the settings in the sample QABot_config.json file.