Provide clearer information if changes not being written to persdata
Although there is a warning at the top of the output file the tool
output for any changes was prefixed with ** irrespective of whether the
JSON files had actually been updated. Now (**) will be used if the flag
is not set, and a separate warning, reminding the user to rerun the tool
with the flag set to TRUE if they want to keep the changes in persdata.