Check for duplicate items with same x,y - not just quoins
Testing revealed that sometimes jellisacs etc are placed close together
and so it is possible that the searching algorithm ends up putting a
jellisac at the same x,y as another. This code will search out these
cases and set the item whose original x,y is furthest away as 'missing'.