Handle closely clustered items more accurately
Previously, when searching with an image, if a perfect match was found
nearby, then that was assumed to be the result and the search stopped.
However that then led to duplicate items being recorded. However it is
necessary to search using all image variants to see if a match is found
closer to the original x,y - thereby producing a better result and
eliminating these false duplicate situations. This code implements this
change for barnacles, ice and jellisacs.