Fix to rewrite duplicate item with the original x,y, variant
The code currently writes out all the changes to file, before then
checking for duplicates and reporting them to the user. However in
addition, these changed duplicates which are now marked as MISSING
should therefore have the changes reversed back to the original
settings.

Really the code should be changed so that the checking for duplicates is
done before we write anything to file. Will looking into making this
change next if it is straightforward.